% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlms.R
\name{all_but_one_labelled}
\alias{all_but_one_labelled}
\title{Check whether all values but one in the middle have labels}
\usage{
all_but_one_labelled(x)
}
\arguments{
\item{x}{labelled vector}
}
\value{
TRUE/FALSE
}
\description{
Check whether all values but one in the middle have labels
}
\details{
Check whether all non-NA but one in the middle values of a labelled variable have labels.
}

