% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlms.R
\name{all_but_rlmsna_labelled}
\alias{all_but_rlmsna_labelled}
\title{Check whether all values but rlms na (99999990+)  have labels}
\usage{
all_but_rlmsna_labelled(x)
}
\arguments{
\item{x}{labelled vector}
}
\value{
TRUE/FALSE
}
\description{
Check whether all values but rlms na (99999990+)  have labels
}
\details{
Check whether all values but rlms na (99999990+)  have labels
}

