% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlms.R
\name{all_labelled}
\alias{all_labelled}
\title{Check whether all values have labels}
\usage{
all_labelled(x, na.rm = TRUE)
}
\arguments{
\item{x}{labelled vector}

\item{na.rm}{a logical value indicating whether NA values should be stripped. TRUE by default}
}
\value{
TRUE/FALSE
}
\description{
Check whether all values have labels
}
\details{
Check whether all values of a labelled variable have labels.
}
\examples{
x <- haven::labelled(c(1, 1, 2, 3, 4), c(Male = 1, Male = 2, Female = 3))
all_labelled(x)
}

