% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlms.R
\name{get_label}
\alias{get_label}
\title{Get variable label}
\usage{
get_label(x)
}
\arguments{
\item{x}{a vector}
}
\value{
character variable label
}
\description{
Get variable label
}
\details{
Get variable label
}

