% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.rlms}
\alias{read.rlms}
\title{Read rlms data}
\usage{
read.rlms(file, suppress = FALSE, nine2na = TRUE)
}
\arguments{
\item{file}{the filename}

\item{suppress}{logical, if true the default message is suppressed}

\item{nine2na}{automatically convert 99999999 to NA for numeric variables}
}
\value{
dataframe
}
\description{
Read rlms data and all meta information
}
\details{
Read rlms data and all the meta information. Destroy useless attributes.
}
\examples{
read.rlms("r21i_os24a.sav")
}

