% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlms.R
\name{rlms_cleanup}
\alias{rlms_cleanup}
\title{Clean rlms data}
\usage{
rlms_cleanup(df, suppress = TRUE, nine2na = TRUE, yesno = TRUE,
  apostrophe = TRUE, remove_empty = TRUE, colnames_tolower = TRUE,
  verbose = FALSE)
}
\arguments{
\item{df}{data.frame read by haven::read.spss}

\item{suppress}{logical, if true the default message is suppressed}

\item{nine2na}{automatically convert 99999999 to NA for numeric variables}

\item{yesno}{convert yes/no answers to lowercase yes/no without apostrophes}

\item{apostrophe}{trim apostrophes, TRUE by default}

\item{remove_empty}{remove empty labels, TRUE by default}

\item{colnames_tolower}{a logical value, indicating whether variable names should be converted to lowercase.}

\item{verbose}{add some debugging output
TRUE by default.}
}
\value{
dataframe
}
\description{
Clean rlms data and all meta information
}
\details{
Clean rlms data and all the meta information. Destroy useless attributes.
}

