% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlms.R
\name{rlms_extract_value_labels}
\alias{rlms_extract_value_labels}
\title{Extract value labels}
\usage{
rlms_extract_value_labels(df)
}
\arguments{
\item{df}{data.frame read from rlms file}
}
\value{
data frame with value labels
}
\description{
Extract value labels
}
\details{
Extract value labels
}

