% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlms.R
\name{rlms_extract_variable_labels}
\alias{rlms_extract_variable_labels}
\title{Extract variable labels}
\usage{
rlms_extract_variable_labels(df)
}
\arguments{
\item{df}{data.frame read from rlms file}
}
\value{
data frame with variable labels
}
\description{
Extract variable labels
}
\details{
Extract variable labels
}

