% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rlms_fileinfo}
\alias{rlms_fileinfo}
\title{Detect wave, sample and level from filename of rlms file}
\usage{
rlms_fileinfo(flist)
}
\arguments{
\item{flist}{the string vector of filenames with or without path}
}
\value{
data.frame containing short filename, wave, level and sample columns
}
\description{
Detect wave, sample and level from filename of rlms file
}
\details{
RLMS filenames contain info about the number of wave, sample (representative or all) and
level (household, individual and a special case of reproductive)
}
\examples{
rlms_fileinfo("r06hall23.sav")
# specify rlms folder first
flist_long <- list.files("~/Downloads/Все выборки/", recursive = TRUE, pattern = "*.sav")
rlms_fileinfo(flist_long)
}

