% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rlms.R
\name{rlms_hints}
\alias{rlms_hints}
\title{Display some RLMS related hints}
\usage{
rlms_hints()
}
\value{
nothing, just prints some messages
}
\description{
Display some RLMS related hints
}
\details{
Display some RLMS related hints
}
\examples{
rlms_hints()
}

