% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlms.R
\name{rlms_legacy_read}
\alias{rlms_legacy_read}
\title{Read rlms data, old legacy code}
\usage{
rlms_legacy_read(file, suppress = TRUE, nine2na = TRUE, yesno = TRUE,
  apostrophe = TRUE, remove_empty = TRUE, colnames_tolower = TRUE)
}
\arguments{
\item{file}{the filename}

\item{suppress}{logical, if true the default message is suppressed}

\item{nine2na}{automatically convert 99999999 to NA for numeric variables}

\item{yesno}{convert yes/no answers to lowercase yes/no without apostrophes}

\item{apostrophe}{trim apostrophes, TRUE by default}

\item{remove_empty}{remove empty labels, TRUE by default}

\item{colnames_tolower}{a logical value, indicating whether variable names should be converted to lowercase.
TRUE by default.}
}
\value{
dataframe
}
\description{
Read rlms data and all meta information, old legacy code
}
\details{
Read rlms data and all the meta information. Destroy useless attributes, old legacy code
}
\examples{
# rlms_legacy_read("r21i_os24a.sav")
}

