% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rlms.R
\name{rlms_load}
\alias{rlms_load}
\title{Load RLMS data of specified wave/level/sample}
\usage{
rlms_load(rlms_folder, wave, level = c("individual", "household",
  "reproductive"), sample = c("all", "representative"))
}
\arguments{
\item{rlms_folder}{path to rlms data}

\item{wave}{the number of wave}

\item{level}{the level (individual/household/reproductive)}

\item{sample}{the sample (all/representative)}
}
\value{
data.frame with RLMS data
}
\description{
Load RLMS data of specified wave/level/sample
}
\details{
Load RLMS data of specified wave/level/sample. This function automatically
determines the file name. The function tries to load .Rds file.
If .Rds file is missing then .sav file is loaded.
}
\examples{
# rlms_load("~/Documents/rlms_data/", wave = 20, level = "individual", sample = "rep" )
}

