% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rlms.R
\name{rlms_read}
\alias{rlms_read}
\title{Read rlms data}
\usage{
rlms_read(file, suppress = FALSE, nine2na = TRUE)
}
\arguments{
\item{file}{the filename}

\item{suppress}{logical, if true the default message is suppressed}

\item{nine2na}{automatically convert 99999999 to NA for numeric variables}
}
\value{
dataframe
}
\description{
Read rlms data and all meta information
}
\details{
Read rlms data and all the meta information. Destroy useless attributes.
}
\examples{
# rlms_read("r21i_os24a.sav")
}

