% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlms.R
\name{rlms_remove_apostrophe}
\alias{rlms_remove_apostrophe}
\title{Remove apostrophes}
\usage{
rlms_remove_apostrophe(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector with apostrophes removed
}
\description{
Remove apostrophes
}
\details{
Remove apostrophes
}

