% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlms.R
\name{rlms_sav2rds}
\alias{rlms_sav2rds}
\title{Convert all RLMS files from .sav to .Rds}
\usage{
rlms_sav2rds(rlms_folder, flatten = TRUE)
}
\arguments{
\item{rlms_folder}{path to rlms data}

\item{flatten}{a logical value indicating whether to flatten folder structure, default is TRUE}
}
\value{
nothing
}
\description{
Convert all RLMS files from .sav to .Rds
}
\details{
Convert all RLMS files from .sav to .Rds
}
\examples{
# rlms_sav2rds("~/Documents/rlms_data/")
}

