% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlms.R
\name{rlms_show_value_labels}
\alias{rlms_show_value_labels}
\title{Show value labels}
\usage{
rlms_show_value_labels(df)
}
\arguments{
\item{df}{data.frame read from rlms file}
}
\value{
data frame with value labels
}
\description{
Show value labels
}
\details{
Show value labels
}

