% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlms.R
\name{rlms_show_variable_labels}
\alias{rlms_show_variable_labels}
\title{Show variable labels}
\usage{
rlms_show_variable_labels(df)
}
\arguments{
\item{df}{data.frame read from rlms file}
}
\value{
data frame with variable labels
}
\description{
Show variable labels
}
\details{
Show variable labels
}

