% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rlms.R
\name{rlms_yesno_standartize}
\alias{rlms_yesno_standartize}
\title{Standartize yes/no}
\usage{
rlms_yesno_standartize(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector with standartized yes/no
}
\description{
Standartize yes/no
}
\details{
Standartize yes/no
}

