% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlms.R
\name{unlabelled_values}
\alias{unlabelled_values}
\title{Get unlabelled values of a labelled vector}
\usage{
unlabelled_values(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector}

\item{na.rm}{a logical value indicating whether NA values should be stripped}
}
\value{
vector of values without labels
}
\description{
Get unlabelled values of a labelled vector
}
\details{
Get unlabelled values of a labelled vector
}

