% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{waves_info}
\alias{waves_info}
\title{Short description of all RLMS waves.}
\format{A data frame with 77 rows and 4 variables:
\describe{
  \item{file_short}{filename}
  \item{wave}{wave number}
  \item{level}{individual, household or reproductive}
  \item{sample}{all or representative}
}}
\source{
\url{https://www.hse.ru/rlms}, 2016-05-31.
}
\usage{
waves_info
}
\description{
Short description of all RLMS waves.
}
\keyword{datasets}

