% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlms.R
\name{all_labelled}
\alias{all_labelled}
\title{Check whether all values have labels}
\usage{
all_labelled(x)
}
\arguments{
\item{x}{labelled vector}
}
\value{
TRUE/FALSE
}
\description{
Check whether all values have labels
}
\details{
Check whether all non-NA values of a labelled variable have labels.
}

