% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlms.R
\name{as_factor_safe}
\alias{as_factor_safe}
\title{Safe version of as_factor.}
\usage{
as_factor_safe(x)
}
\arguments{
\item{x}{labelled vector}
}
\value{
TRUE/FALSE
}
\description{
Safe version of as_factor.
}
\details{
Safe version of as_factor. This function keeps unlabelled values and 
does not replace them with NA as `as_factor` in `haven` package.
It also avoids "duplicated levels" warning.
}
\examples{
x <- haven::labelled(c(1, 1, 2, 3, 4), c(Male = 1, Male = 2, Female = 3))
as_factor_safe(x)
}

