% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlms.R
\name{is_labelled}
\alias{is_labelled}
\title{Check whether the variable is labelled}
\usage{
is_labelled(x)
}
\arguments{
\item{x}{a vector}
}
\value{
TRUE/FALSE
}
\description{
Check whether the variable is labelled
}
\details{
Check whether the variable is labelled
}
\examples{
x <- haven::labelled(c(1, 1, 2, 3, 4), c(Male = 1, Male = 2, Female = 3))
is_labelled(x)
}

