% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlms.R
\name{labelled_values}
\alias{labelled_values}
\title{Get labelled values of a labelled vector}
\usage{
labelled_values(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector}

\item{na.rm}{a logical value indicating whether NA values should be stripped.
Normally NA is not labelled and is not returned even with na.rm = FALSE.}
}
\value{
vector of values with labels
}
\description{
Get labelled values of a labelled vector
}
\details{
Get labelled values of a labelled vector
}

