% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlms.R
\name{read.rlms}
\alias{read.rlms}
\title{Read rlms data (deprecated function)}
\usage{
read.rlms(file, suppress = FALSE, nine2na = TRUE)
}
\arguments{
\item{file}{the filename}

\item{suppress}{logical, if true the default message is suppressed}

\item{nine2na}{automatically convert 99999999 to NA for numeric variables}
}
\value{
dataframe
}
\description{
Read rlms data and all meta information (deprecated function)
}
\details{
Read rlms data and all the meta information. Destroy useless attributes.
This function is deprecated. Please use rlms_read instead of.
}
\examples{
# read.rlms("r21i_os24a.sav")
}

