% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rlms_hints}
\alias{rlms_hints}
\title{Display some RLMS related hints}
\usage{
rlms_hints()
}
\value{
nothing, just prints some messages
}
\description{
Display some RLMS related hints
}
\details{
Display some RLMS related hints
}
\examples{
rlms_hints()
}

