% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlms.R
\name{rlms_read}
\alias{rlms_read}
\title{Read rlms data}
\usage{
rlms_read(file, haven = c("factor", "labelled", "numeric"), suppress = TRUE,
  ...)
}
\arguments{
\item{file}{the filename}

\item{haven}{use haven package:
"labelled" - return labelled variables,
"factor" - return factor or numeric variables,
"numeric" - return numeric variables.}

\item{suppress}{logical, if true the default message is suppressed}

\item{...}{further parameters passed to rlms_cleanup() function}
}
\value{
dataframe
}
\description{
Read rlms data and all meta information
}
\details{
Read rlms data and all the meta information. Destroy useless attributes.
}
\examples{
# rlms_read("r21i_os24a.sav")
}

