% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlms.R
\name{rlms_read}
\alias{rlms_read}
\title{Read rlms data}
\usage{
rlms_read(file, suppress = TRUE, nine2na = TRUE, yesno = TRUE,
  apostrophe = TRUE, remove_empty = TRUE, haven = c("no", "labelled",
  "factor", "numeric"))
}
\arguments{
\item{file}{the filename}

\item{suppress}{logical, if true the default message is suppressed}

\item{nine2na}{automatically convert 99999999 to NA for numeric variables}

\item{yesno}{convert yes/no answers to lowercase yes/no without apostrophes}

\item{apostrophe}{trim apostrophes, TRUE by default}

\item{remove_empty}{remove empty labels, TRUE by default}

\item{haven}{use haven package: "no" - (default) do not use "haven" package,
"labelled" - return labelled variables,
"factor" - return factor or numeric variables,
"numeric" - return numeric variables.}
}
\value{
dataframe
}
\description{
Read rlms data and all meta information
}
\details{
Read rlms data and all the meta information. Destroy useless attributes.
}
\examples{
# rlms_read("r21i_os24a.sav")
}

